+++ jopa-impl/src/main/java/cz/cvut/kbss/jopa/query/sparql/loader/UnboundPredicateObjectSparqlAssemblyModifier.java
@@ -1,7 +1,10 @@
-package cz.cvut.kbss.jopa.query.sparql;
+package cz.cvut.kbss.jopa.query.sparql.loader;
 
 import cz.cvut.kbss.jopa.query.QueryParameter;
 import cz.cvut.kbss.jopa.query.QueryType;
+import cz.cvut.kbss.jopa.query.sparql.QueryAttributes;
+import cz.cvut.kbss.jopa.query.sparql.TokenQueryParameter;
+import cz.cvut.kbss.jopa.query.sparql.TokenStreamSparqlQueryHolder;
 import org.antlr.v4.runtime.Token;
 import org.antlr.v4.runtime.TokenStream;
 import org.antlr.v4.runtime.TokenStreamRewriter;
@@ -11,9 +14,9 @@ import org.antlr.v4.runtime.TokenStreamRewriter;
  * <p>
  * This optimizer is applicable for SELECT queries that select instances of an entity class. Instead of loading the
  * instances one by one after the query is evaluated, this optimizer modifies the query to fetch all available entity
- * attributes.
+ * attributes by injecting a triple pattern with unbound predicate and object variables into the query.
  */
-public class EntityLoadingSparqlAssemblyModifier implements SparqlAssemblyModifier {
+public class UnboundPredicateObjectSparqlAssemblyModifier implements SparqlAssemblyModifier {
 
     @Override
     public void modify(TokenStreamSparqlQueryHolder queryHolder, TokenStreamRewriter tokenRewriter,
@@ -22,7 +25,7 @@ public class EntityLoadingSparqlAssemblyModifier implements SparqlAssemblyModifi
         assert queryHolder.getProjectedQueryParameters().size() == 1;
 
         final TokenQueryParameter<?> p = queryHolder.getProjectedQueryParameters().get(0);
-        tokenRewriter.insertAfter(p.getTokens().get(0), generateProjectionModification(p));
+        tokenRewriter.insertAfter(p.getSingleToken(), generateProjectionModification(p));
         tokenRewriter.insertBefore(queryAttributes.lastClosingCurlyBraceToken(), generateSelectionTriplePattern(p, tokenRewriter.getTokenStream(), queryAttributes.lastClosingCurlyBraceToken()));
     }
 
@@ -31,7 +34,7 @@ public class EntityLoadingSparqlAssemblyModifier implements SparqlAssemblyModifi
         return " " + property(baseName) + " " + value(baseName);
     }
 
-    private static String getBaseParamName(QueryParameter<?> p) {
+    static String getBaseParamName(QueryParameter<?> p) {
         return p.getName() != null ? p.getName() : p.getPosition().toString();
     }
 
@@ -44,7 +47,7 @@ public class EntityLoadingSparqlAssemblyModifier implements SparqlAssemblyModifi
     }
 
     private static String generateSelectionTriplePattern(TokenQueryParameter<?> subjectParam, TokenStream tokenStream,
-                                                  Token lastClosingCurlyBracket) {
+                                                         Token lastClosingCurlyBracket) {
         final String baseName;
         final char varPrefix;
         if (subjectParam.getPosition() != null) {