+++ jopa-impl/src/test/java/cz/cvut/kbss/jopa/query/sparql/loader/UnboundPredicateObjectSparqlAssemblyModifierTest.java
@@ -1,31 +1,22 @@
-package cz.cvut.kbss.jopa.query.sparql;
+package cz.cvut.kbss.jopa.query.sparql.loader;
 
 import cz.cvut.kbss.jopa.query.parameter.ParameterValueFactory;
+import cz.cvut.kbss.jopa.query.sparql.Sparql11QueryParser;
+import cz.cvut.kbss.jopa.query.sparql.TokenStreamSparqlQueryHolder;
 import cz.cvut.kbss.jopa.sessions.MetamodelProvider;
-import cz.cvut.kbss.jopa.sessions.UnitOfWork;
-import cz.cvut.kbss.jopa.utils.Configuration;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
 import static org.hamcrest.MatcherAssert.assertThat;
 import static org.hamcrest.Matchers.equalToCompressingWhiteSpace;
 import static org.mockito.Mockito.mock;
-import static org.mockito.Mockito.when;
 
-class EntityLoadingSparqlAssemblyModifierTest {
+class UnboundPredicateObjectSparqlAssemblyModifierTest {
 
     private final ParameterValueFactory valueFactory = new ParameterValueFactory(mock(MetamodelProvider.class));
 
-    private Sparql11QueryParser parser;
+    private final Sparql11QueryParser parser = new Sparql11QueryParser(valueFactory);
 
-    private final EntityLoadingSparqlAssemblyModifier sut = new EntityLoadingSparqlAssemblyModifier();
-
-    @BeforeEach
-    void setUp() {
-        final UnitOfWork uowMock = mock(UnitOfWork.class);
-        when(uowMock.getConfiguration()).thenReturn(new Configuration());
-        this.parser = new Sparql11QueryParser(valueFactory);
-    }
+    private final UnboundPredicateObjectSparqlAssemblyModifier sut = new UnboundPredicateObjectSparqlAssemblyModifier();
 
     @Test
     void modifyAddsUnboundPropertyAndValuePatternAndProjectsPropertyAndValueForNamedParameter() {